backend_django/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
drf-spectacular==0.27.2
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
